+++ 07-problems-level-3/p02.cpp
@@ -26,7 +26,7 @@ void printMatrix(int matrix[3][3], short matrixRows, short matrixColumns, string
 	cout << prompt;
 	for (short i = 0; i < matrixRows; i++)
 	{
-		for (int j = 0; j < matrixColumns; j++)
+		for (short j = 0; j < matrixColumns; j++)
 		{
 			cout << setw(3) << matrix[i][j] << "\t";
 		}
@@ -38,7 +38,7 @@ void fillMatrixWithRandomNumbers(int matrix[3][3],short matrixRows, short matrix
 {
 	for (short i = 0; i < matrixRows; i++)
 	{
-		for (int j = 0; j < matrixColumns; j++)
+		for (short j = 0; j < matrixColumns; j++)
 		{
 			matrix[i][j] = getRandomNumber(1, 100);
 		}
